+++ CHANGELOG.md
@@ -6,7 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v0.2.1
+
+### Added
+
+- `Map::get_mut` and `Map::get_field_mut` provide exclusive access to the fields
+  that are found.
+- `Set::remove_member` removes a member by index.
+- `Set` now implements `Debug`, `Ord`, `Eq`, and `Clone`.
+
+## v0.2.0
 
 ### Added